Move project fixture higher up to use in view testing
